+++ app/src/main/java/com/tripplanner/previous_trip_details/previousTripDetailsFragment.java
@@ -89,8 +89,7 @@ public class previousTripDetailsFragment extends Fragment implements OnMapReadyC
         RecyclerView.LayoutManager mLayoutManager = new LinearLayoutManager(getContext());
         mViewModel = ViewModelProviders.of(this).get(PreviousTripDetailsViewModel.class);
         // TODO: Use the ViewModel
-
-        noteAdapter =new NoteAdapter(mViewModel.getTripNotes((int)trip.getId()));
+        noteAdapter =new NoteAdapter(mViewModel.getTripNotes(trip.getId()));
         noteRecycler.setAdapter(noteAdapter);
         noteRecycler.setLayoutManager(mLayoutManager);
